backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7  # Required for PyMySQL with SSL
alembic==1.12.1  # Database migrations

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP requests and async
httpx==0.25.2
aiofiles==23.2.1

# Validation and serialization
pydantic[email]==2.5.0
pydantic-settings==2.0.3

# Environment and config
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Rate limiting and caching
slowapi==0.1.9
redis==5.0.1  # Optional for advanced rate limiting

# Payment gateways
requests==2.31.0  # For payment API calls

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2  # For testing

# Production
uvloop==0.19.0  # Performance improvement for Linux
